mostly kinda done

+++ keepr.client/src/components/KeepDetails.vue
@@ -1,14 +1,14 @@
 <template>
   <div v-if="keep && accountVaults" class="row justify-content-evenly align-content-center">
     <div class="col-6 pm-0">
-      <img class="img-fluid img-card rounded" :src="keep.img" alt="">
+      <img class="img-fluid img-card rounded" :title="keep.name" :src="keep.img" alt="">
     </div>
     <div class="col-6 d-flex flex-column justify-content-evenly">
       <div class="row justify-content-evenly">
         <div class="d-flex justify-content-center">
-          <i class="mdi mdi-eye mx-2"></i>
+          <i class="mdi mdi-eye mx-2" title="Keep Views"></i>
           <p>{{ keep.views }}</p>
-          <i class="mdi mdi-alpha-k-box-outline mx-2"></i>
+          <i class="mdi mdi-alpha-k-box-outline mx-2" title="Kept Count"></i>
           <p>{{ keep.kept }}</p>
         </div>
       </div>
@@ -21,21 +21,23 @@
       <div class="row">
         <div class="col-6 d-flex justify-content-start align-items-center">
           <div class="mx-5">
-            <select v-model="editable.id" v-if="account" name="vaults" id="vaults">
-              <option v-for="v in accountVaults" :value="v.id">{{ v.name }}</option>
+            <select v-model="editable.id" title="Vaults" v-if="account" name="vaults" id="vaults">
+              <option v-for="v in accountVaults" :title="v.name" :value="v.id">{{ v.name }}</option>
             </select>
-            <button class="btn btn-primary" @click="saveToVault(editable, keep.id)">Save</button>
+            <button class="btn btn-primary" title="Save to Vault Keep"
+              @click="saveToVault(editable, keep.id)">Save</button>
           </div>
         </div>
         <div class="col-6 d-flex justify-content-end">
-          <div v-if="keep.creatorId == account.id" class="d-flex fs-4 mx-2 selectable" @click="deleteKeep(keep.id)">
+          <div v-if="keep.creatorId == account.id" class="d-flex fs-4 mx-2 selectable" title="Delete Keep"
+            @click="deleteKeep(keep.id)">
             <i class="mdi mdi-delete d-flex align-items-center"></i>
             <h6 class="d-flex align-items-center text-center m-0">Delete</h6>
           </div>
           <div>
             <router-link :to="{ name: 'ProfilePage', params: { profileId: keep.creatorId } }">
-              <img class="img-modal rounded-circle img-fluid mx-2" :src="keep.creator.picture" alt=""
-                data-bs-dismiss="modal">
+              <img class="img-modal rounded-circle img-fluid mx-2" :title="keep.creator.name"
+                :src="keep.creator.picture" alt="" data-bs-dismiss="modal">
             </router-link>
           </div>
           <div class="d-flex align-items-center">
@@ -59,6 +61,7 @@ import { Modal } from "bootstrap";
 import { profilesService } from "../services/ProfilesService.js";
 import { vaultKeepsService } from "../services/VaultKeepsService.js";
 import { accountService } from "../services/AccountService.js";
+import { useRoute } from "vue-router";
 export default {
   props: {
     keep: {
@@ -67,16 +70,20 @@ export default {
   },
   setup() {
     const editable = ref({});
+    const route = useRoute;
     return {
       editable,
+      route,
       account: computed(() => AppState.account),
       profileVaults: computed(() => AppState.profileVaults),
       activeVault: computed(() => AppState.activeVault),
       accountVaults: computed(() => AppState.accountVaults),
       async deleteKeep(keepId) {
         try {
-          await keepsService.deleteKeep(keepId);
-          Modal.getOrCreateInstance('#keepDetails').hide();
+          if (await Pop.confirm("are you sure you want to delete this keep?", "this could do some mad damage yo")) {
+            await keepsService.deleteKeep(keepId);
+            Modal.getOrCreateInstance('#keepDetails').hide();
+          }
         } catch (error) {
           logger.error(error);
           Pop.error(error.message);
@@ -93,7 +100,7 @@ export default {
           logger.error(error);
           Pop.error(error.message);
         }
-      }
+      },
     };
   }
 };
